+++ src/routes/Navigation.svelte
@@ -35,6 +35,7 @@
     <li><a href="/tests/array-component">Array proxy</a></li>
     <li><a href="/tests/redirect-same-route">Redirect with delay</a></li>
     <li><a href="/tests/strict-mode">Strict mode</a></li>
+    <li><a href="/tests/submit-enter">Submit with enter</a></li>
   </ul>
 </nav>