+++ src/routes/tests/submit-enter/+page.server.ts
@@ -1,0 +1,23 @@
+import { superValidate, setError } from '$lib/server/index.js';
+import { schema } from './schema.js';
+
+import type { Actions, PageServerLoad } from './$types.js';
+
+///// Load function /////
+
+export const load: PageServerLoad = async () => {
+  const form = await superValidate(schema);
+  return { form };
+};
+
+///// Form actions /////
+
+export const actions: Actions = {
+  default: async ({ request }) => {
+    const form = await superValidate(request, schema);
+
+    console.log('POST', form);
+
+    return setError(form, 'name', 'some error');
+  }
+};